Moved apply_custom_command and made it Public
