+++ wlx-overlay-s/src/gui/panel/mod.rs
@@ -1,5 +1,6 @@
 use std::{cell::RefCell, rc::Rc};
 
+use anyhow::Context;
 use button::setup_custom_button;
 use glam::{Affine2, Vec2, vec2};
 use label::setup_custom_label;
@@ -8,15 +9,23 @@ use wgui::{
     components::button::ComponentButton,
     drawing,
     event::{
-        Event as WguiEvent, EventCallback, EventListenerID, EventListenerKind,
-        InternalStateChangeEvent, MouseButtonIndex, MouseDownEvent, MouseLeaveEvent,
-        MouseMotionEvent, MouseUpEvent, MouseWheelEvent,
+        CallbackDataCommon, Event as WguiEvent, EventAlterables, EventCallback, EventListenerID,
+        EventListenerKind, InternalStateChangeEvent, MouseButtonIndex, MouseDownEvent,
+        MouseLeaveEvent, MouseMotionEvent, MouseUpEvent, MouseWheelEvent,
     },
     gfx::cmd::WGfxClearMode,
+    i18n::Translation,
     layout::{Layout, LayoutParams, LayoutUpdateParams, WidgetID},
-    parser::{self, CustomAttribsInfoOwned, Fetchable, ParseDocumentExtra, ParserState},
+    parser::{
+        self, CustomAttribsInfoOwned, Fetchable, ParseDocumentExtra, ParserState, parse_color_hex,
+    },
     renderer_vk::context::Context as WguiContext,
-    widget::{EventResult, label::WidgetLabel},
+    renderer_vk::text::custom_glyph::CustomGlyphData,
+    taffy,
+    widget::{
+        EventResult, image::WidgetImage, label::WidgetLabel, rectangle::WidgetRectangle,
+        sprite::WidgetSprite,
+    },
     windowing::context_menu::{self, ContextMenu},
 };
 use wlx_common::overlays::{BackendAttrib, BackendAttribValue};
@@ -25,6 +34,7 @@ use wlx_common::timestep::Timestep;
 use crate::{
     app_misc,
     backend::input::{Haptics, HoverResult, PointerHit, PointerMode},
+    backend::task::ModifyPanelCommand,
     state::AppState,
     subsystem::hid::WheelDelta,
     windowing::backend::{
@@ -473,4 +483,103 @@ fn log_cmd_invalid_arg(
         "{:?}: <{tag_name}> \"{attrib}\": \"{command}\" has invalid argument: {arg}",
         parser_state.path.get_path_buf()
     )
+}
+
+pub fn apply_custom_command<T>(
+    panel: &mut GuiPanel<T>,
+    app: &mut AppState,
+    element: &str,
+    command: &ModifyPanelCommand,
+) -> anyhow::Result<()> {
+    let mut alterables = EventAlterables::default();
+    let mut com = CallbackDataCommon {
+        alterables: &mut alterables,
+        state: &panel.layout.state,
+    };
+
+    match command {
+        ModifyPanelCommand::SetText(text) => {
+            if let Ok(mut label) = panel
+                .parser_state
+                .fetch_widget_as::<WidgetLabel>(&panel.layout.state, element)
+            {
+                label.set_text(&mut com, Translation::from_raw_text(text));
+            } else if let Ok(button) = panel
+                .parser_state
+                .fetch_component_as::<ComponentButton>(element)
+            {
+                button.set_text(&mut com, Translation::from_raw_text(text));
+            } else {
+                anyhow::bail!("No <label> or <Button> with such id.");
+            }
+        }
+        ModifyPanelCommand::SetImage(path) => {
+            if let Ok(pair) = panel
+                .parser_state
+                .fetch_widget(&panel.layout.state, element)
+            {
+                let data = CustomGlyphData::from_assets(
+                    &app.wgui_globals,
+                    wgui::assets::AssetPath::File(path),
+                )
+                .context("Could not load content from supplied path.")?;
+
+                if let Some(mut sprite) = pair.widget.get_as::<WidgetSprite>() {
+                    sprite.set_content(&mut com, Some(data));
+                } else if let Some(mut image) = pair.widget.get_as::<WidgetImage>() {
+                    image.set_content(&mut com, Some(data));
+                } else {
+                    anyhow::bail!("No <sprite> or <image> with such id.");
+                }
+            } else {
+                anyhow::bail!("No <sprite> or <image> with such id.");
+            }
+        }
+        ModifyPanelCommand::SetColor(color) => {
+            let color = parse_color_hex(color)
+                .context("Invalid color format, must be a html hex color!")?;
+
+            if let Ok(pair) = panel
+                .parser_state
+                .fetch_widget(&panel.layout.state, element)
+            {
+                if let Some(mut rect) = pair.widget.get_as::<WidgetRectangle>() {
+                    rect.set_color(&mut com, color);
+                } else if let Some(mut label) = pair.widget.get_as::<WidgetLabel>() {
+                    label.set_color(&mut com, color, true);
+                } else if let Some(mut sprite) = pair.widget.get_as::<WidgetSprite>() {
+                    sprite.set_color(&mut com, color);
+                } else {
+                    anyhow::bail!("No <rectangle> or <label> or <sprite> with such id.");
+                }
+            } else {
+                anyhow::bail!("No <rectangle> or <label> or <sprite> with such id.");
+            }
+        }
+        ModifyPanelCommand::SetVisible(visible) => {
+            let wid = panel
+                .parser_state
+                .get_widget_id(element)
+                .context("No widget with such id.")?;
+
+            let display = if *visible {
+                taffy::Display::Flex
+            } else {
+                taffy::Display::None
+            };
+
+            com.alterables
+                .set_style(wid, wgui::event::StyleSetRequest::Display(display));
+        }
+        ModifyPanelCommand::SetStickyState(sticky_down) => {
+            let button = panel
+                .parser_state
+                .fetch_component_as::<ComponentButton>(element)
+                .context("No <Button> with such id.")?;
+            button.set_sticky_state(&mut com, *sticky_down);
+        }
+    }
+
+    panel.layout.process_alterables(alterables)?;
+    Ok(())
 }
