+++ wlx-overlay-s/src/overlays/watch.rs
@@ -1,20 +1,13 @@
 use std::{rc::Rc, time::Duration};
 
-use anyhow::Context;
 use glam::{Affine3A, Quat, Vec3, Vec3A, vec3};
 use wgui::{
     assets::AssetPath,
     components::button::ComponentButton,
     event::{CallbackDataCommon, EventAlterables, StyleSetRequest},
     layout::WidgetID,
-    parser::{Fetchable, ParseDocumentParams,parse_color_hex},
-    renderer_vk::text::custom_glyph::CustomGlyphData,
+    parser::{Fetchable, ParseDocumentParams},
     taffy,
-    widget::{
-        image::WidgetImage, label::WidgetLabel, rectangle::WidgetRectangle, sprite::WidgetSprite,
-    },    
-    i18n::Translation,
-
 };
 
 use wlx_common::{
@@ -23,11 +16,10 @@ use wlx_common::{
 };
 
 use crate::{
-        backend::task::ModifyPanelCommand,
     gui::{
         panel::{
-            GuiPanel, NewGuiPanelParams, device_list::DeviceList, overlay_list::OverlayList,
-            set_list::SetList,
+            GuiPanel, NewGuiPanelParams, apply_custom_command, device_list::DeviceList,
+            overlay_list::OverlayList, set_list::SetList,
         },
         timer::GuiTimer,
     },
@@ -84,86 +76,84 @@ pub fn create_watch(app: &mut AppState) -> anyhow::Result<OverlayWindowConfig> {
     };
 
     panel.on_notify = Some(Box::new({
-        let name=WATCH_NAME;
+        let name = WATCH_NAME;
         move |panel, app, event_data| {
-        let mut alterables = EventAlterables::default();
+            let mut alterables = EventAlterables::default();
 
-        let mut elems_changed = panel.state.overlay_list.on_notify(
-            &mut panel.layout,
-            &mut panel.parser_state,
-            &event_data,
-            &mut alterables,
-            &doc_params,
-        )?;
+            let mut elems_changed = panel.state.overlay_list.on_notify(
+                &mut panel.layout,
+                &mut panel.parser_state,
+                &event_data,
+                &mut alterables,
+                &doc_params,
+            )?;
 
-        elems_changed |= panel.state.set_list.on_notify(
-            &mut panel.layout,
-            &mut panel.parser_state,
-            &event_data,
-            &mut alterables,
-            &doc_params,
-        )?;
+            elems_changed |= panel.state.set_list.on_notify(
+                &mut panel.layout,
+                &mut panel.parser_state,
+                &event_data,
+                &mut alterables,
+                &doc_params,
+            )?;
 
-        elems_changed |= panel.state.device_list.on_notify(
-            app,
-            &mut panel.layout,
-            &mut panel.parser_state,
-            &event_data,
-            &doc_params,
-        )?;
+            elems_changed |= panel.state.device_list.on_notify(
+                app,
+                &mut panel.layout,
+                &mut panel.parser_state,
+                &event_data,
+                &doc_params,
+            )?;
 
-
-
-        match event_data {
-            OverlayEventData::EditModeChanged(edit_mode) => {
-                if let Ok(btn_edit_mode) = panel
-                    .parser_state
-                    .fetch_component_as::<ComponentButton>("btn_edit_mode")
-                {
-                    let mut com = CallbackDataCommon {
-                        alterables: &mut alterables,
-                        state: &panel.layout.state,
-                    };
-                    btn_edit_mode.set_sticky_state(&mut com, edit_mode);
+            match event_data {
+                OverlayEventData::EditModeChanged(edit_mode) => {
+                    if let Ok(btn_edit_mode) = panel
+                        .parser_state
+                        .fetch_component_as::<ComponentButton>("btn_edit_mode")
+                    {
+                        let mut com = CallbackDataCommon {
+                            alterables: &mut alterables,
+                            state: &panel.layout.state,
+                        };
+                        btn_edit_mode.set_sticky_state(&mut com, edit_mode);
+                    }
                 }
-            }
-            OverlayEventData::SettingsChanged => {
-                panel.layout.mark_redraw();
-                sets_or_overlays(panel, app, &mut alterables);
+                OverlayEventData::SettingsChanged => {
+                    panel.layout.mark_redraw();
+                    sets_or_overlays(panel, app, &mut alterables);
 
-                if app.session.config.clock_12h != panel.state.clock_12h {
-                    panel.state.clock_12h = app.session.config.clock_12h;
+                    if app.session.config.clock_12h != panel.state.clock_12h {
+                        panel.state.clock_12h = app.session.config.clock_12h;
 
-                    let clock_root = panel.parser_state.get_widget_id("clock_root")?;
-                    panel.layout.remove_children(clock_root);
+                        let clock_root = panel.parser_state.get_widget_id("clock_root")?;
+                        panel.layout.remove_children(clock_root);
 
-                    panel.parser_state.instantiate_template(
-                        &doc_params,
-                        "Clock",
-                        &mut panel.layout,
-                        clock_root,
-                        Default::default(),
-                    )?;
+                        panel.parser_state.instantiate_template(
+                            &doc_params,
+                            "Clock",
+                            &mut panel.layout,
+                            clock_root,
+                            Default::default(),
+                        )?;
 
-                    elems_changed = true;
+                        elems_changed = true;
+                    }
                 }
-            }
-            OverlayEventData::CustomCommand {element, command} =>{
-                if let Err(e) = apply_custom_command(panel, app, &element, &command) {
-                    log::warn!("Could not apply {command:?} on {name}/{element}: {e:?}");
+                OverlayEventData::CustomCommand { element, command } => {
+                    if let Err(e) = apply_custom_command(panel, app, &element, &command) {
+                        log::warn!("Could not apply {command:?} on {name}/{element}: {e:?}");
+                    }
                 }
+                _ => {}
             }
-            _ => {}
+
+            if elems_changed {
+                panel.process_custom_elems(app);
+            }
+
+            panel.layout.process_alterables(alterables)?;
+            Ok(())
         }
-
-
-        if elems_changed {
-            panel.process_custom_elems(app);
-        }
-
-        panel.layout.process_alterables(alterables)?;
-        Ok(())
-    }}));
+    }));
 
     panel
         .timers
@@ -237,102 +227,3 @@ pub fn watch_fade<D>(app: &mut AppState, watch: &mut OverlayWindowData<D>) {
     state.alpha += 0.1;
     state.alpha = state.alpha.clamp(0., 1.);
 }
-fn apply_custom_command(
-    panel: &mut GuiPanel<WatchState>,
-    app: &mut AppState,
-    element: &str,
-    command: &ModifyPanelCommand,
-) -> anyhow::Result<()> {
-    let mut alterables = EventAlterables::default();
-    let mut com = CallbackDataCommon {
-        alterables: &mut alterables,
-        state: &panel.layout.state,
-    };
-
-    match command {
-        ModifyPanelCommand::SetText(text) => {
-            if let Ok(mut label) = panel
-                .parser_state
-                .fetch_widget_as::<WidgetLabel>(&panel.layout.state, element)
-            {
-                label.set_text(&mut com, Translation::from_raw_text(text));
-            } else if let Ok(button) = panel
-                .parser_state
-                .fetch_component_as::<ComponentButton>(element)
-            {
-                button.set_text(&mut com, Translation::from_raw_text(text));
-            } else {
-                anyhow::bail!("No <label> or <Button> with such id.");
-            }
-        }
-        ModifyPanelCommand::SetImage(path) => {
-            if let Ok(pair) = panel
-                .parser_state
-                .fetch_widget(&panel.layout.state, element)
-            {
-                let data = CustomGlyphData::from_assets(
-                    &app.wgui_globals,
-                    wgui::assets::AssetPath::File(path),
-                )
-                .context("Could not load content from supplied path.")?;
-
-                if let Some(mut sprite) = pair.widget.get_as::<WidgetSprite>() {
-                    sprite.set_content(&mut com, Some(data));
-                } else if let Some(mut image) = pair.widget.get_as::<WidgetImage>() {
-                    image.set_content(&mut com, Some(data));
-                } else {
-                    anyhow::bail!("No <sprite> or <image> with such id.");
-                }
-            } else {
-                anyhow::bail!("No <sprite> or <image> with such id.");
-            }
-        }
-        ModifyPanelCommand::SetColor(color) => {
-            let color = parse_color_hex(color)
-                .context("Invalid color format, must be a html hex color!")?;
-
-            if let Ok(pair) = panel
-                .parser_state
-                .fetch_widget(&panel.layout.state, element)
-            {
-                if let Some(mut rect) = pair.widget.get_as::<WidgetRectangle>() {
-                    rect.set_color(&mut com, color);
-                } else if let Some(mut label) = pair.widget.get_as::<WidgetLabel>() {
-                    label.set_color(&mut com, color, true);
-                } else if let Some(mut sprite) = pair.widget.get_as::<WidgetSprite>() {
-                    sprite.set_color(&mut com, color);
-                } else {
-                    anyhow::bail!("No <rectangle> or <label> or <sprite> with such id.");
-                }
-            } else {
-                anyhow::bail!("No <rectangle> or <label> or <sprite> with such id.");
-            }
-        }
-        ModifyPanelCommand::SetVisible(visible) => {
-            let wid = panel
-                .parser_state
-                .get_widget_id(element)
-                .context("No widget with such id.")?;
-
-            let display = if *visible {
-                taffy::Display::Flex
-            } else {
-                taffy::Display::None
-            };
-
-            com.alterables
-                .set_style(wid, wgui::event::StyleSetRequest::Display(display));
-        }
-        ModifyPanelCommand::SetStickyState(sticky_down) => {
-            let button = panel
-                .parser_state
-                .fetch_component_as::<ComponentButton>(element)
-                .context("No <Button> with such id.")?;
-            button.set_sticky_state(&mut com, *sticky_down);
-        }
-    }
-
-    panel.layout.process_alterables(alterables)?;
-    Ok(())
-}
-
